Recyclerviews are added to show notifications, in this case a friend request in activity_notifications.xml.
In ApiService.java the endpoint /get_requests is added to be used in NotificationsActivity.java in the loadFriendRequests method.
In FriendRequest.java getter strings of getEmail_amigo and optional setters are added.

In RetrofitClient.java a method is created to obtain an Apiclient instance.

In SearchFriendsActivity the logic for the navigation bar is added

3 new files are created: item_friend_request.xml to show the requests to accept or not accept it and its logic FriendRequestAdapter.java